refactor: extract webhook domain into ferriskey-webhook
