+++ core/Cargo.toml
@@ -16,6 +16,7 @@ ferriskey-mail = { path = "../libs/ferriskey-mail" }
 ferriskey-compass = { path = "../libs/ferriskey-compass" }
 ferriskey-seawatch = { path = "../libs/ferriskey-seawatch", features = ["mock"] }
 ferriskey-password-policy = { path = "../libs/ferriskey-password-policy", features = ["mock"] }
+ferriskey-webhook = { path = "../libs/ferriskey-webhook", features = ["mock"] }
 ferriskey-migrate = { path = "../libs/ferriskey-migrate" }
 maskass = { path = "../libs/maskass" }
 anyhow = "1.0.98"
