+++ libs/ferriskey-webhook/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "ferriskey-webhook"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-domain = { path = "../ferriskey-domain" }
+chrono = { version = "0.4.41", features = ["serde"] }
+serde = { version = "1.0.219", features = ["derive"] }
+thiserror = "2.0.12"
+utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
+uuid = { version = "1.16.0", features = ["serde", "v4", "v7"] }
+mockall = { version = "0.14.0", optional = true }
+
+[dev-dependencies]
+mockall = "0.14.0"
+
+[features]
+mock = ["mockall"]
