+++ libs/ferriskey-webhook/src/entities.rs
