+++ libs/ferriskey-webhook/src/entities/errors.rs
